initial work on conductor rust implementation
this is going to be an exonum app (rust-based blockchain). this is a
natural use because the point of conductor is a distributed database
with products/hiring shared across regions.

so far the "Accounts" model is fairly built out (which is the basic for
any currency) and now starting to build out companies as well.

there is no api at this point, and no (known) method for publishing
transactions from within conductor itself. i would *like* to keep the
entire app contained to this rust repo if possible. if the entire system
could be run from one compiled binary, that would be ideal. however,
it's likely that eventually (or even early on) some kind of app server
which uses conductor's blockchain as a db would make sense. still
mulling this over, and honestly, it doesn't matter yet because the db
isn't even built yet.