now including all orders int he rolling index
changing the rolling index for orders to INCLUDE non-finalized orders,
and index them by create date. i am realizing using only the
update+finalized orders is too rigid (an order cannot be marked as
anything other than finalized after or it would corrupt the index and
double-dount the order once it was finalized again). using create date
or non-finalized orders is not ideal, but it's more "resilient" and
also doesn't prevent us from doing filtering when pulling the orders
out. the only real problem would be orders that take longer than a year
to fill, but we can cross that bridge when we get there (likely by
increasing the time window on a per-company basis).

this has one side effect, which is actually what spurred the change in
the first place: if we're trying to come up with costs for a product, we
can include the current non-finalized orders in our calculations. what
we do is say if we have less than 10 finalized orders, include
non-finalized orders in the calculations as well. this helps alleviate
bloated costs for the first few orders.

in other words, if i make electricity and it's sold by the watthour and
someone makes an order of 10000 watthours, even if i've only worked one
day, the cost previously would be 8 * 10000, or 80000 labor hours for
the order (because the outgoing costs of the order are not included
because the order isn't finalized yet!) by including the order itself as
a cost divider (remember, the basic algorithm is inputs/outputs) then
the cost of a watthour of electricity becomes (8 / 10000) instead (and
at 10000wh, the total cost for the order is 8 labor hours, which is
pretty much exactly what we want).

this is a good change.