limiting places where costs are updated
the idea is to only update product costs when absolutely necessary. for
instance, we were doing it when a labor record is finalized, and doing
it for the orderer when an order is created.

now, we only update costs for the company receiving the order when
either the order is created or the status is updated.

hoping this gives us some more scaling headroom.