Implementing *tracking* of cost aggregates
Cost aggregates are now tracked and updated as orders are rolled into/out
of the rolling indexes. This means that when calculating costs, we no
longer need to pull out all the orders that affect costs, but rather
just pull out the aggregate buckets (one lookup vs potentially 1000s).

Note that we haven't implemented using the aggregates in the cost algo
yet, however the tracking is live.

This change should save potentially huge amounts of resources when
calculating costs for orders and hopefully will scale magnitudes better
than before. The only tricky part is it adds a layer of complexity that
might take a few tries to get right. Some more tests are definitely in
order here.

Note that because of this, we're also ripping out the 1-per-day
calculation cap on costs.