fully implementing cost tags
god damn, this was a big one.

- gutted the costing algorithm, moved the code responsible for tallying
costs to a new Costable trait in the cost_tag model
- cost algo now fully uses cost tags (ripped out Effort, sum_hours,
product Inputs, etc etc...all the automated junk is gone now).
- implemented Costable for labor/orders, and used the new trait in the
aggregate processing in the schema AND the raw order processing of the
cost algo (meaning we no longer duplicate the code for tallying up costs
of various objects)
- allowing cost tag assignments for members (as `default_cost_tags`),
labor, orders, and products
- making sure only existing/active cost tags can be assigned
- updating company model such that cost tags can be created when the
company is created, AND assigned to the founder's member record at the
same time.
- updating company model such that cost tags/founder are sub-objects
when creating a new company (sorry, integration tests)
- updated all tests, passing!! (the main worries are cost/tallying
tests)

just need to "deal" with integration tests now.